# swagger/api.yaml
swagger: "2.0"
info:
  title: "BitShares Explorer API"
  description: "Exposes BitShares calls needed to explore the blockchain"
  version: "2.0.0"
schemes:
  - http
  - https
produces:
  - application/json
paths:
  /block/blockNum:
    get:
      summary: Retrieve block data
      description: Retrieve block data by block number
      parameters:
        - in: path
          name: blockNum
          required: true
          schema:
            type: integer
          description: Numeric ID of the block to retrieve
      responses:
        "200":
          description: Block data
        "500":
          description: Server error
  /txs/{blockNum}:
    get:
      summary: Retrieve transactions for a block
      description: Retrieve transactions for a block by block number
      parameters:
        - in: path
          name: blockNum
          required: true
          schema:
            type: integer
          description: Numeric ID of the block to retrieve transactions from
      responses:
        "200":
          description: List of transactions
        "500":
          description: Server error
  /account/accountCount:
    get:
      summary: Total number of registered accounts
      description: Get the total number of accounts registered with the blockchain
      responses:
        "200":
          description: Accounts count
        "500":
          description: Server error
  /account/{accountName}:
    get:
      summary: Retrieve account data
      description: Retrieve account data by account name
      parameters:
        - in: path
          name: accountName
          required: true
          schema:
            type: string
          description: Name of the account to retrieve
      responses:
        "200":
          description: Account data
        "500":
          description: Server error
  /account/{accountRef}:
    get:
      summary: Retrieve account references
      description: Retrieve references of accounts by account reference
      parameters:
        - in: path
          name: accountRef
          required: true
          schema:
            type: string
          description: Reference of the account to retrieve references for
      responses:
        "200":
          description: Account references
        "500":
          description: Server error
  /chainID:
    get:
      summary: Retrieve chain ID
      description: Retrieve the chain ID of the blockchain
      responses:
        "200":
          description: Chain ID
        "500":
          description: Server error
  block/latestBlock:
    get:
      summary: Retrieve the latest block data
      description: Retrieve the data of the latest block on the blockchain
      responses:
        "200":
          description: Latest block data
        "500":
          description: Server error
  /tx/{transaction}:
    get:
      summary: Retrieve transaction data
      description: Retrieve transaction data by transaction hash
      parameters:
        - in: path
          name: transaction
          required: true
          schema:
            type: string
          description: Hash of the transaction to retrieve
      responses:
        "200":
          description: Transaction data
        "500":
          description: Server error